refactor: spring cleaning
- Remove redundant unused code
- Cleanup the way `Link` is used
- Fix some styles
- simplify meta tags handling by relying the route hierarchy!
