--- app/components/ArrowIcon.tsx
@@ -1,9 +1,0 @@
-import { FaArrowLeft, FaArrowRight } from "react-icons/fa";
-
-export const ArrowRightIcon = () => (
-  <FaArrowRight className="align-middle inline" size={12} />
-);
-
-export const ArrowLeftIcon = () => (
-  <FaArrowLeft className="align-middle inline" size={12} />
-);
