--- app/components/ExternalLink.tsx
@@ -1,22 +1,0 @@
-import React from "react";
-
-export const ExternalLink = ({
-  href,
-  children,
-  className = "",
-}: {
-  href: string;
-  children: React.ReactNode;
-  className?: string;
-}) => {
-  return (
-    <a
-      href={href}
-      target="_blank"
-      rel="noreferrer"
-      className={`primary-link ${className}`}
-    >
-      {children}
-    </a>
-  );
-};
