+++ app/components/SiteHeader.tsx
@@ -1,5 +1,4 @@
-import { Link } from "@remix-run/react";
-import { ExternalLink } from "./ExternalLink";
+import { ExternalLink, InternalLink } from "./Link";
 import { Separator } from "./Separator";
 
 export const SiteHeader = () => {
@@ -7,21 +6,21 @@ export const SiteHeader = () => {
     <div className="prose mb-3 w-full max-w-full print:hidden">
       <h2 className="mb-2">Hello, I'm Krishna 🧑🏽‍💻</h2>
       <p className="my-3">
-        <Link className="primary-link mr-4" to="/">
+        <InternalLink className="mr-4" to="/">
           About
-        </Link>
-        <Link className="primary-link mr-4" to="/resume">
+        </InternalLink>
+        <InternalLink className="mr-4" to="/resume">
           Résumé
-        </Link>
+        </InternalLink>
         <ExternalLink
           href="https://github.com/krishnagopinath/website/raw/master/files/resume.pdf"
-          className="primary-link mr-4"
+          className="mr-4"
         >
           Résumé (PDF)
         </ExternalLink>
-        <Link className="primary-link mr-4" to="/posts">
-          Blog
-        </Link>
+        <InternalLink className="mr-4" to="/posts">
+          Writing
+        </InternalLink>
       </p>
       <Separator />
     </div>
