+++ app/components/resume/Footer.tsx
