+++ app/components/resume/Skills.tsx
@@ -1,6 +1,6 @@
 import type { ResumeSchema } from "@kurone-kito/jsonresume-types";
 import type { ValuesType } from "utility-types";
-import { List, ListItem } from "./List";
+import { List, ListItem } from "../List";
 import { SubHeading } from "./SubHeading";
 
 type SkillEntries = NonNullable<ResumeSchema["skills"]>;
