+++ app/components/resume/WorkExperience.tsx
@@ -1,8 +1,8 @@
 import type { ResumeSchema } from "@kurone-kito/jsonresume-types";
 import type { ValuesType } from "utility-types";
-import { ExternalLink } from "~/components/ExternalLink";
+import { ExternalLink } from "~/components/Link";
 import { formatDate, Formats } from "~/utils/dates";
-import { List, ListItem } from "./List";
+import { List, ListItem } from "../List";
 import { SubHeading } from "./SubHeading";
 
 type WorkExperienceEntries = NonNullable<ResumeSchema["work"]>;
