+++ app/data/about.mdx
@@ -1,5 +1,5 @@
 import { SocialLinks } from "~/components/SocialLinks";
-import { ExternalLink } from "~/components/ExternalLink";
+import { ExternalLink, InternalLink } from "~/components/Link";
 
 I'm a **software engineer**, **budding teacher** and an **amateur photographer**.
 
@@ -15,6 +15,7 @@ My work principles are to strive to be _iterative_, _communicative_ and _thoroug
 During my free time, I could be doing any of the following:
 
 - Rewriting <ExternalLink href="https://github.com/Matterwiki/Matterwiki/tree/refactor">Matterwiki</ExternalLink>, a simple wiki for small teams.
+- Trying to build a habit of <InternalLink to="/posts">writing</InternalLink>.
 - Capturing <ExternalLink href="https://www.instagram.com/krishwader">pictures of food, places and architecture</ExternalLink>.
 - Cooking vegetarian recipes from a multitude of cuisines.
