+++ app/routes/index.tsx
@@ -1,8 +1,14 @@
-import { metaTagsBuilder } from "~/utils/metaTagsBuilder";
 import type { MetaFunction } from "@remix-run/node";
 import About from "~/data/about.mdx";
 
-export const meta: MetaFunction = () => metaTagsBuilder("Home");
+export const meta: MetaFunction = () => {
+  const title = `Home | krishnagopinath.me`;
+  return {
+    title,
+    "og:title": title,
+    "twitter:title": title,
+  };
+};
 
 export default function AboutPage() {
   return (
