+++ app/routes/posts.tsx
@@ -1,4 +1,14 @@
+import type { MetaFunction } from "@remix-run/node";
 import { Outlet } from "@remix-run/react";
+
+export const meta: MetaFunction = () => {
+  const title = `Writing | krishnagopinath.me`;
+  return {
+    title,
+    "og:title": title,
+    "twitter:title": title,
+  };
+};
 
 export default function PostsPage() {
   return (
