+++ app/routes/posts/hello-world.mdx
@@ -1,10 +1,14 @@
 ---
-title: Hello world
-description: First blog post on krishnagopinath.me
-draft: false
-date: 2022-05-08
+meta:
+  title: Hello world
+  og:title: Hello world
+  twitter:title: Hello world
+  description: First blog post on krishnagopinath.me
+  og:description": First blog post on krishnagopinath.me
+  twitter:description": First blog post on krishnagopinath.me
+  draft: false
+  date: 2022-05-08
 ---
-
 # Hello world
 
 Hey, welcome to my blog 🙌
@@ -14,4 +18,5 @@ This is a `hello world` post, if you will. I plan to write here about the follow
 * Technology
 * Software Engineering
 * Photography
-* Cooking
+* Food & Cooking
+* Productivity
